Add files via upload
Uploading several files to explore options for our ML model.  The Test.ipynb is using linear regression to predict prices.  The other three are different examples of using classification models.  The Classification ML.ipynb file used binning for prices by evenly distributing them into four bins.  The accuracy of these models were not desirable.  The last two notebooks instead was using classification models to predict which county a house is in.  The notebook with lat-long contains the models that had upwards of 99% accuracy, but that was to be expected when we are trying to predict county using coordinates.